RW-4589 update styles
